+++ src/renderer/src/components/codex-restart-notice-key.ts
@@ -1,0 +1,6 @@
+export function buildCodexRestartNoticeKey(args: {
+  previousAccountLabel: string
+  nextAccountLabel: string
+}): string {
+  return `${args.previousAccountLabel}\u0000${args.nextAccountLabel}`
+}
